v3.2.0 | Display Booking Logs
This version adds options to display the booking logs on either the plugin settings page, the Dashboard or a static Page.
On the static page a summary of the logs is displayed to all showing a breakdown of booking by pickup point and if the user is logged in as an Admin, additionally the full booking detailed list.